requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
langchain-chroma==0.2.0
openai==1.54.0 
faiss-cpu==1.7.4
chromadb==0.5.3
sentence-transformers==3.0.1
pypdf==4.2.0
python-docx==1.1.2
markdown==3.6
pandas==2.2.2
fastapi==0.112.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
redis==5.0.8
fastapi-cache2  
python-dotenv==1.0.1